Better support 1.17
Shortened prefix a hair, rely on display name instead of name if name appears to be truncated. This allows us up to 128 characters on 1.17.

Always truncate display name - seems like an oversight on Spigot's end that they allow naming to 32767 characters but limit the display name to 128 characters. As display name internally uses team name until set, this allows bypassing the display name limit. As we automatically generate team names, we should be mindful of potentially illegal states.

Closes #12